requirements.txt
phidata
python-dotenv
yfinance
duckduckgo-search
fastapi
uvicorn
groq
openai
packaging
newspaper4k 
lxml_html_clean
streamlit
watchdog